+++ cogs/meme/meme.py
@@ -8,9 +8,13 @@ from typing import List, Union
 import discord
 from discord.ext import commands
 
+from core.config import config
 from cogs.resource import CogConfig, CogText
-from core import rubbercog, image_utils, utils
+from core import image_utils, rubbercog, utils
 from core.emote import emote
+from repository.interaction_repo import InteractionRepository
+
+repo_i = InteractionRepository()
 
 
 class Meme(rubbercog.Rubbercog):
@@ -24,6 +28,7 @@ class Meme(rubbercog.Rubbercog):
 
         self.fishing_pool = self.config.get("_fishing")
 
+    @commands.guild_only()
     @commands.cooldown(rate=5, per=20.0, type=commands.BucketType.user)
     @commands.command()
     async def hug(self, ctx, user: discord.Member = None):
@@ -32,13 +37,46 @@ class Meme(rubbercog.Rubbercog):
         user: Discord user. If none, the bot will hug yourself.
         """
         if user is None:
-            user = ctx.author
-        elif user == self.bot.user:
-            await ctx.send(emote.hug_left)
-            return
+            hugger = self.bot.user
+            hugged = ctx.author
+        else:
+            hugger = ctx.author
+            hugged = user
 
-        await ctx.send(emote.hug_right + f" **{self.sanitise(user.display_name)}**")
+        repo_i.add(ctx.guild.id, ctx.channel.id, ctx.message.id, "hug", hugger.id, hugged.id)
+        await ctx.send(f"{emote.hug_right} **{self.sanitise(hugged.display_name)}**")
 
+    @commands.guild_only()
+    @commands.cooldown(rate=5, per=20.0, type=commands.BucketType.user)
+    @commands.command()
+    async def whip(self, ctx, user: discord.Member = None):
+        """Whip someone"""
+        if user is None:
+            whipper = self.bot.user
+            whipped = ctx.author
+        else:
+            whipper = ctx.author
+            whipped = user
+
+        repo_i.add(ctx.guild.id, ctx.channel.id, ctx.message.id, "whip", whipper.id, whipped.id)
+        await ctx.send(f"{self.config.get('whip')} **{self.sanitise(whipped.display_name)}**")
+
+    @commands.guild_only()
+    @commands.cooldown(rate=5, per=20.0, type=commands.BucketType.user)
+    @commands.command()
+    async def spank(self, ctx, user: discord.Member = None):
+        """Spank someone"""
+        if user is None:
+            spanker = self.bot.user
+            spanked = ctx.author
+        else:
+            spanker = ctx.author
+            spanked = user
+
+        repo_i.add(ctx.guild.id, ctx.channel.id, ctx.message.id, "spank", spanker.id, spanked.id)
+        await ctx.send(f"{self.config.get('spank')} **{self.sanitise(spanked.display_name)}**")
+
+    @commands.guild_only()
     @commands.cooldown(rate=5, per=20.0, type=commands.BucketType.user)
     @commands.command()
     async def pet(self, ctx, member: discord.Member = None):
@@ -47,10 +85,16 @@ class Meme(rubbercog.Rubbercog):
         member: Discord user. If none, the bot will hug yourself.
         """
         if member is None:
-            member = ctx.author
+            petter = self.bot.user
+            petted = ctx.author
+        else:
+            petter = ctx.author
+            petted = member
+
+        repo_i.add(ctx.guild.id, ctx.channel.id, ctx.message.id, "pet", petter.id, petted.id)
 
         async with ctx.typing():
-            url = member.avatar_url_as(format="jpg")
+            url = petted.avatar_url_as(format="jpg")
             response = requests.get(url)
             avatar = Image.open(BytesIO(response.content))
 
@@ -69,9 +113,11 @@ class Meme(rubbercog.Rubbercog):
                     optimize=False,
                 )
                 image_binary.seek(0)
-                filename = self.get_pet_name(member)
+
+                filename = self.get_pet_name(petted)
                 await ctx.send(file=discord.File(fp=image_binary, filename=filename))
 
+    @commands.guild_only()
     @commands.cooldown(rate=5, per=20.0, type=commands.BucketType.user)
     @commands.command()
     async def hyperpet(self, ctx, member: discord.Member = None):
@@ -80,10 +126,16 @@ class Meme(rubbercog.Rubbercog):
         member: Discord user. If none, the bot will hug yourself.
         """
         if member is None:
-            member = ctx.author
+            petter = self.bot.user
+            petted = ctx.author
+        else:
+            petter = ctx.author
+            petted = member
+
+        repo_i.add(ctx.guild.id, ctx.channel.id, ctx.message.id, "hyperpet", petter.id, petted.id)
 
         async with ctx.typing():
-            url = member.avatar_url_as(format="jpg")
+            url = petted.avatar_url_as(format="jpg")
             response = requests.get(url)
             avatar = Image.open(BytesIO(response.content))
 
@@ -102,8 +154,52 @@ class Meme(rubbercog.Rubbercog):
                     optimize=False,
                 )
                 image_binary.seek(0)
-                filename = self.get_pet_name(member)
+                filename = self.get_pet_name(petted)
                 await ctx.send(file=discord.File(fp=image_binary, filename=filename))
+
+    @commands.guild_only()
+    @commands.cooldown(rate=5, per=20.0, type=commands.BucketType.user)
+    @commands.command()
+    async def slap(self, ctx, member: discord.Member = None):
+        """Slap someone!
+
+        member: Discord user. If none, the bot will slap yourself.
+        """
+        if member is None:
+            slapper = self.bot.user
+            slapped = ctx.author
+        else:
+            slapper = ctx.author
+            slapped = member
+
+        options = ["つ", "づ", "ノ"]
+
+        repo_i.add(ctx.guild.id, ctx.channel.id, ctx.message.id, "slap", slapper.id, slapped.id)
+
+        await ctx.send(
+            "**{}**{} {}".format(
+                self.sanitise(slapper.display_name),
+                random.choice(options),
+                self.sanitise(slapped.display_name),
+            )
+        )
+
+    @commands.guild_only()
+    @commands.cooldown(rate=1, per=5, type=commands.BucketType.user)
+    @commands.command()
+    async def relations(self, ctx, user: discord.User = None):
+        """Get your information about hugs, pets, ..."""
+        if user is None:
+            user = ctx.author
+
+        embed = self.embed(ctx=ctx, description=self.sanitise(user.display_name))
+
+        for action in ("hug", "pet", "hyperpet", "slap", "spank", "whip"):
+            lookup = repo_i.get_user_action(user.id, ctx.guild.id, action)
+            value = self.text.get("value", gave=lookup[0], got=lookup[1])
+            embed.add_field(name=f"{config.prefix}{action}", value=value)
+
+        await ctx.send(embed=embed)
 
     @commands.cooldown(rate=5, per=120, type=commands.BucketType.user)
     @commands.command(aliases=["owo"])