+++ resources/backup.standalone.sh
@@ -9,9 +9,9 @@ pg_dump rubbergoddess > dump_`date +%Y-%m-%d"_"%H:%M:%S`.sql
 
 today=$(date +%d)
 
-if [ $today -eq "1" ]; then
+if [ $today -eq "01" ]; then
 	# compress last month
-	month=$(date -d "`date +%Y%m01` -1day" +%Y_%m)
+	month=$(date -d "`date +%Y%m01` -1day" +%Y-%m)
 	tar -cJf dump_$month.tar.xz dump_$month*
 	rm dump_$month*
 fi